botapp/requirements.txt
celery==4.1.0
Flask==1.0.2
gunicorn==19.8.1
pymongo==3.6.1
python-telegram-bot==10.1.0
redis==2.10.6
requests==2.18.4
python-dotenv==0.8.2